move atari network to examples/atari/network.py
