+++ test/base/test_utils.py
@@ -4,7 +4,6 @@ import numpy as np
 from tianshou.utils import MovAvg
 from tianshou.utils import SummaryWriter
 from tianshou.utils.net.common import MLP, Net
-from tianshou.utils.net.discrete import DQN, C51
 from tianshou.exploration import GaussianNoise, OUNoise
 from tianshou.utils.net.continuous import RecurrentActorProb, RecurrentCritic
 
@@ -76,17 +75,6 @@ def test_net():
     data = torch.rand([bsz, 8, np.prod(state_shape)])
     act = torch.rand(expect_output_shape)
     assert list(net(data, act).shape) == [bsz, 1]
-    # DQN
-    state_shape = (4, 84, 84)
-    action_shape = (6, )
-    data = np.random.rand(bsz, *state_shape)
-    expect_output_shape = [bsz, *action_shape]
-    net = DQN(*state_shape, action_shape)
-    assert list(net(data)[0].shape) == expect_output_shape
-    num_atoms = 51
-    net = C51(*state_shape, action_shape, num_atoms)
-    expect_output_shape = [bsz, *action_shape, num_atoms]
-    assert list(net(data)[0].shape) == expect_output_shape
 
 
 def test_summary_writer():
